# Use tinyerc55 to validate EthereumAddress