--- a/packages/shared-pure/package.json
+++ b/packages/shared-pure/package.json
@@ -16,6 +16,7 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
+    "@mradomski/tinyerc55": "^0.1.0",
     "chalk": "^4.1.2",
     "ethers": "^5.7.2",
     "zod": "^3.22.2"