--- a/packages/shared-pure/package.json
+++ b/packages/shared-pure/package.json
@@ -16,7 +16,7 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@mradomski/tinyerc55": "^0.1.0",
+    "@mradomski/tinyerc55": "^0.1.1",
     "chalk": "^4.1.2",
     "ethers": "^5.7.2",
     "zod": "^3.22.2"