Tested with `hyperfine --warmup 5 "echo echo 'require("@l2beat/config")' | node"`

Before:
```
Benchmark 1: ethers
  Time (mean ± σ):      1.442 s ±  0.036 s    [User: 1.588 s, System: 0.208 s]
  Range (min … max):    1.413 s …  1.528 s    10 runs
```

After:
```
Benchmark 1: tinyerc55
  Time (mean ± σ):     975.7 ms ±   9.9 ms    [User: 1137.3 ms, System: 192.4 ms]
  Range (min … max):   966.3 ms … 1000.0 ms    10 runs
```